+++ common/network/components/router/router_model.h
@@ -15,7 +15,7 @@ class NetPacket;
 class RouterModel
 {
 public:
-   RouterModel(NetworkModel* model, float frequency,
+   RouterModel(NetworkModel* model, double frequency,
                SInt32 num_input_ports, SInt32 num_output_ports,
                SInt32 num_flits_per_port_buffer, UInt64 delay, SInt32 flit_width,
                bool contention_model_enabled, string& contention_model_type);
@@ -48,7 +48,7 @@ public:
 
 private:
    NetworkModel* _model;
-   volatile float _frequency;
+   volatile double _frequency;
    SInt32 _num_input_ports;
    SInt32 _num_output_ports;
    SInt32 _flit_width;
