+++ common/tile/memory_subsystem/cache/cache.cc
@@ -22,8 +22,8 @@ Cache::Cache(string name,
              CacheReplacementPolicy* replacement_policy,
              CacheHashFn* hash_fn,
              UInt32 access_delay,
-             float frequency,
-             float voltage,
+             double frequency,
+             double voltage,
              bool track_miss_types)
    : _enabled(false)
    , _name(name)
