+++ common/tile/memory_subsystem/memory_manager.h
@@ -16,7 +16,7 @@ void MemoryManagerNetworkCallback(void* obj, NetPacket packet);
 class MemoryManager
 {
 public:
-   MemoryManager(Tile* tile, float frequency, float voltage);
+   MemoryManager(Tile* tile, double frequency, double voltage);
    virtual ~MemoryManager();
 
    bool __coreInitiateMemoryAccess(MemComponent::Type mem_component,
@@ -48,7 +48,7 @@ public:
    virtual bool isModeled(const void* pkt_data) = 0;
 
    static CachingProtocolType parseProtocolType(std::string& protocol_type);
-   static MemoryManager* createMMU(std::string protocol_type, Tile* tile, float frequency, float voltage);
+   static MemoryManager* createMMU(std::string protocol_type, Tile* tile, double frequency, double voltage);
    
    virtual void computeEnergy() = 0;
