+++ common/tile/memory_subsystem/pr_l1_pr_l2_dram_directory_mosi/dram_directory_cntlr.h
@@ -35,8 +35,8 @@ namespace PrL1PrL2DramDirectoryMOSI
                          string dram_directory_type_str,
                          UInt32 num_dram_cntlrs,
                          string dram_directory_access_time_str,
-                         float frequency,
-                         float voltage);
+                         double frequency,
+                         double voltage);
       ~DramDirectoryCntlr();
 
       void handleMsgFromL2Cache(tile_id_t sender, ShmemMsg* shmem_msg);
