+++ common/tile/memory_subsystem/pr_l1_pr_l2_dram_directory_mosi/memory_manager.h
@@ -25,7 +25,7 @@ namespace PrL1PrL2DramDirectoryMOSI
    class MemoryManager : public ::MemoryManager
    {
    public:
-      MemoryManager(Tile* tile, float frequency, float voltage);
+      MemoryManager(Tile* tile, double frequency, double voltage);
       ~MemoryManager();
 
       UInt32 getCacheLineSize() { return _cache_line_size; }
