+++ common/tile/memory_subsystem/pr_l1_pr_l2_dram_directory_msi/l2_cache_cntlr.cc
@@ -18,8 +18,8 @@ L2CacheCntlr::L2CacheCntlr(MemoryManager* memory_manager,
                            string l2_cache_replacement_policy,
                            UInt32 l2_cache_access_delay,
                            bool l2_cache_track_miss_types,
-                           float frequency,
-                           float voltage)
+                           double frequency,
+                           double voltage)
    : _memory_manager(memory_manager)
    , _l1_cache_cntlr(l1_cache_cntlr)
    , _dram_directory_home_lookup(dram_directory_home_lookup)
