+++ common/tile/memory_subsystem/pr_l1_pr_l2_dram_directory_msi/memory_manager.cc
@@ -9,7 +9,7 @@
 namespace PrL1PrL2DramDirectoryMSI
 {
 
-MemoryManager::MemoryManager(Tile* tile, float frequency, float voltage)
+MemoryManager::MemoryManager(Tile* tile, double frequency, double voltage)
    : ::MemoryManager(tile, frequency, voltage)
    , _dram_directory_cntlr(NULL)
    , _dram_cntlr(NULL)
