+++ common/tile/memory_subsystem/pr_l1_sh_l2_msi/memory_manager.cc
@@ -10,7 +10,7 @@
 namespace PrL1ShL2MSI
 {
 
-MemoryManager::MemoryManager(Tile* tile, float frequency, float voltage)
+MemoryManager::MemoryManager(Tile* tile, double frequency, double voltage)
    : ::MemoryManager(tile, frequency, voltage)
    , _dram_cntlr(NULL)
    , _dram_cntlr_present(false)
